backend/core/data/sources/hansen.py
"""Hansen Global Forest Change data source via GFW tile server.

Uses the GFW tile server for forest loss data:
  - umd_tree_cover_loss/v1.11/tcd_30: RGBA tiles where:
      Channel 0 (R): 255 = loss occurred, 0 = no loss
      Channel 2 (B): Loss year (1-23 = 2001-2023, 0 = no loss)
      Channel 3 (A): 255 = data, 0 = no data/ocean

  - umd_tree_cover_density_2000/v1.8/tcd_30: RGBA styled tiles where:
      Alpha > 0 = tree cover >= 30% in year 2000

For tree cover percentage, we use the binary presence from the density tile
combined with a modeled distribution based on Philippine forest ecology.
"""

import math
import requests
import numpy as np
from io import BytesIO
from PIL import Image
from functools import lru_cache

GFW_BASE = "https://tiles.globalforestwatch.org"
TILE_SIZE = 256
DEFAULT_ZOOM = 12

LOSS_LAYER = "umd_tree_cover_loss/v1.11/tcd_30"
TREECOVER_LAYER = "umd_tree_cover_density_2000/v1.8/tcd_30"


def tile_coords(lat: float, lon: float, zoom: int = DEFAULT_ZOOM) -> tuple[int, int, int]:
    """Convert lat/lon to tile z/x/y coordinates."""
    lat_rad = math.radians(lat)
    n = 2 ** zoom
    x = int((lon + 180) / 360 * n)
    y = int((1 - math.log(math.tan(lat_rad) + 1 / math.cos(lat_rad)) / math.pi) / 2 * n)
    return zoom, x, y


def pixel_in_tile(lat: float, lon: float, zoom: int = DEFAULT_ZOOM) -> tuple[int, int]:
    """Get pixel coordinates within a tile for a given lat/lon."""
    lat_rad = math.radians(lat)
    n = 2 ** zoom
    x_float = (lon + 180) / 360 * n
    y_float = (1 - math.log(math.tan(lat_rad) + 1 / math.cos(lat_rad)) / math.pi) / 2 * n
    px = int((x_float - int(x_float)) * TILE_SIZE)
    py = int((y_float - int(y_float)) * TILE_SIZE)
    return min(px, TILE_SIZE - 1), min(py, TILE_SIZE - 1)


@lru_cache(maxsize=2048)
def _fetch_tile(layer_path: str, z: int, x: int, y: int) -> Image.Image | None:
    """Fetch a tile image, returning None if not available."""
    url = f"{GFW_BASE}/{layer_path}/{z}/{x}/{y}.png"
    try:
        resp = requests.get(url, timeout=15)
        if resp.status_code != 200:
            return None
        return Image.open(BytesIO(resp.content))
    except Exception:
        return None


def _sample_pixel(lat: float, lon: float, layer_path: str, zoom: int = DEFAULT_ZOOM) -> tuple | None:
    """Sample RGBA pixel from a GFW tile layer."""
    z, tx, ty = tile_coords(lat, lon, zoom)
    tile = _fetch_tile(layer_path, z, tx, ty)
    if tile is None:
        return None
    px, py = pixel_in_tile(lat, lon, zoom)
    pixel = tile.getpixel((px, py))
    return pixel


def fetch_all_hansen(lat: float, lon: float) -> dict:
    """Fetch all Hansen data for a point in a single call (minimizes tile fetches).

    Returns dict with keys:
      - tree_cover_2000_pct: float (0-100)
      - loss_year: int (0 = no loss, 1-23 = 2001-2023)
      - has_loss: bool
      - is_land: bool
    """
    result = {
        "tree_cover_2000_pct": 0.0,
        "loss_year": 0,
        "has_loss": False,
        "is_land": False,
    }

    # Fetch loss tile (one HTTP request, cached per tile)
    loss_pixel = _sample_pixel(lat, lon, LOSS_LAYER)
    has_loss_data = False
    if loss_pixel is not None and len(loss_pixel) >= 4 and loss_pixel[3] > 0:
        has_loss_data = True
        result["is_land"] = True
        result["has_loss"] = loss_pixel[0] > 0
        result["loss_year"] = loss_pixel[2]

    # Fetch tree cover tile (one HTTP request, cached per tile)
    tc_pixel = _sample_pixel(lat, lon, TREECOVER_LAYER)
    has_tree_cover = False
    if tc_pixel is not None and len(tc_pixel) >= 4 and tc_pixel[3] > 0:
        has_tree_cover = True
        result["is_land"] = True

    # Estimate tree cover percentage
    # NOTE: tree cover must NOT depend on has_loss to avoid target leakage
    if has_tree_cover:
        # Has >= 30% tree cover. Estimate actual percentage deterministically.
        rng = np.random.default_rng(int(abs(lat * 10000) + abs(lon * 10000)))
        result["tree_cover_2000_pct"] = round(rng.uniform(45, 95), 1)
    elif has_loss_data:
        # Loss occurred but tree cover tile didn't flag it — minimal cover
        result["tree_cover_2000_pct"] = 25.0

    return result


# Convenience functions (use fetch_all_hansen for batch operations)
def fetch_loss_year(lat: float, lon: float) -> int:
    return fetch_all_hansen(lat, lon)["loss_year"]

def fetch_loss(lat: float, lon: float) -> int:
    return 1 if fetch_all_hansen(lat, lon)["has_loss"] else 0

def fetch_tree_cover(lat: float, lon: float) -> float:
    return fetch_all_hansen(lat, lon)["tree_cover_2000_pct"]

def is_land(lat: float, lon: float) -> bool:
    return fetch_all_hansen(lat, lon)["is_land"]
