backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.2
shap==0.46.0
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2
openai>=1.30.0
pydantic==2.9.2
aiosqlite==0.20.0
geopy==2.4.1
Pillow>=10.0.0
# srtm not available for Python 3.11; using HTTP-based elevation
shapely>=2.0.0
requests>=2.31.0
scipy>=1.11.0
pytest>=8.0.0
xgboost>=2.0.0
